3dLandscapeModel/requirements.txt
freetype

PyOpenGL
pygame

